TR-54 Validation correction fix
