+++ src/main/java/ru/hh/techradar/dto/BlipTemplateDto.java
@@ -1,40 +1,33 @@
 package ru.hh.techradar.dto;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
-import jakarta.validation.constraints.NotBlank;
-import jakarta.validation.constraints.NotNull;
 
 @JsonInclude(JsonInclude.Include.NON_NULL)
 public class BlipTemplateDto {
-
-  @NotBlank(message = "Name must contain at least one symbol!")
   private String name;
-
-  @NotNull(message = "Description must be not null!")
   private String description;
 
-  public BlipTemplateDto(@NotBlank String name, @NotNull String description) {
+  public BlipTemplateDto(String name, String description) {
     this.name = name;
     this.description = description;
   }
 
   public BlipTemplateDto() {
-    description = "";
   }
 
-  public @NotBlank String getName() {
+  public String getName() {
     return name;
   }
 
-  public void setName(@NotBlank String name) {
+  public void setName(String name) {
     this.name = name;
   }
 
-  public @NotNull String getDescription() {
+  public String getDescription() {
     return description;
   }
 
-  public void setDescription(@NotNull String description) {
+  public void setDescription(String description) {
     this.description = description;
   }
 }
